requirements.txt
asgiref==3.4.1
Django==3.2.5
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
psycopg2-binary>=2.8